refactor(print-format-builder): extract SectionPropertiesPanel
