+++ frappe/public/js/print_format_builder/components/inspector/FieldInspector.vue
@@ -589,188 +589,7 @@
 		</template>
 
 		<!-- ── Section inspector ───────────────────────────────── -->
-		<template v-else-if="selected_section">
-			<div class="pfb-insp-body">
-				<!-- SECTION properties -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_section')">
-						<span class="pfb-insp-section-label">{{ __("Section") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_section }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_section" class="pfb-insp-section-body">
-						<!-- Title -->
-						<LabelField
-							v-model="selected_section.label"
-							:label="__('Title')"
-							:placeholder="__('Untitled section')"
-							show-toggle
-							:show-label="__('Show title')"
-							:show="selected_section.show_label"
-							@update:show="(v) => (selected_section.show_label = v)"
-						/>
-
-						<!-- Columns -->
-						<SegmentedRow
-							:label="__('Columns')"
-							:model-value="selected_section.columns.length"
-							:options="[1, 2, 3, 4].map((n) => ({ value: n, label: n }))"
-							@update:model-value="set_columns"
-						/>
-
-						<!-- Orientation -->
-						<SegmentedRow
-							:label="__('Label side')"
-							:model-value="
-								section_orientation === 'left-right' ? 'left-right' : 'top'
-							"
-							:options="[
-								{ value: 'top', label: __('Top') },
-								{ value: 'left-right', label: __('Left') },
-							]"
-							@update:model-value="
-								(v) =>
-									(selected_section.field_orientation =
-										v === 'left-right' ? 'left-right' : '')
-							"
-						/>
-
-						<!-- Gap -->
-						<StepperRow
-							:label="__('Gap')"
-							:model-value="section_gap"
-							:step="4"
-							:base="20"
-							unit="px"
-							@update:model-value="(v) => (selected_section.gap = v)"
-						/>
-					</div>
-				</div>
-
-				<!-- BACKGROUND -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_bg')">
-						<span class="pfb-insp-section-label">{{ __("Background") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_bg }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_bg" class="pfb-insp-section-body">
-						<div ref="bg_color_host"></div>
-					</div>
-				</div>
-
-				<!-- PADDING -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_padding')">
-						<span class="pfb-insp-section-label">{{ __("Padding") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_padding }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_padding" class="pfb-insp-section-body">
-						<PaddingGrid
-							:model-value="selected_section.padding"
-							@update:model-value="(v) => (selected_section.padding = v)"
-						/>
-					</div>
-				</div>
-
-				<!-- LAYOUT -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_layout')">
-						<span class="pfb-insp-section-label">{{ __("Layout") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_layout }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_layout" class="pfb-insp-section-body">
-						<!-- Layout mode -->
-						<SegmentedRow
-							:label="__('Mode')"
-							:model-value="section_field_borders"
-							:options="[
-								{ value: false, label: __('Normal') },
-								{ value: true, label: __('Table') },
-							]"
-							@update:model-value="toggle_field_borders"
-						/>
-						<!-- Cell padding -->
-						<StepperRow
-							:label="__('Cell padding')"
-							:model-value="section_cell_padding"
-							:base="8"
-							unit="px"
-							@update:model-value="(v) => (selected_section.cell_padding = v)"
-						/>
-					</div>
-				</div>
-
-				<!-- STYLE -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_style')">
-						<span class="pfb-insp-section-label">{{ __("Style") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_style }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_style">
-						<StyleSection v-model="selected_section.custom_style" />
-					</div>
-				</div>
-
-				<!-- VISIBILITY -->
-				<div class="pfb-insp-section">
-					<div class="pfb-insp-section-head" @click="toggle('s_visibility')">
-						<span class="pfb-insp-section-label">{{ __("Visibility") }}</span>
-						<span
-							class="pfb-insp-chevron"
-							:class="{ collapsed: !open.s_visibility }"
-							v-html="frappe.utils.icon('chevron-down', 'xs')"
-						></span>
-					</div>
-					<div v-show="open.s_visibility">
-						<VisibilitySection
-							v-model="selected_section.visible_if"
-							:previewDoc="preview_doc"
-						/>
-					</div>
-				</div>
-
-				<div class="pfb-insp-actions">
-					<button
-						class="btn btn-xs btn-danger-subtle"
-						@click="
-							const section = store.selected_section.value;
-							const idx = layout.value.sections.indexOf(section);
-							if (idx !== -1) layout.value.sections.splice(idx, 1);
-							store.selected_section.value = null;
-							if (
-								section &&
-								section.columns.some((c) =>
-									c.fields.includes(store.selected_field.value)
-								)
-							)
-								store.selected_field.value = null;
-						"
-					>
-						<span v-html="frappe.utils.icon('x', 'xs')"></span>
-						{{ __("Remove section") }}
-					</button>
-				</div>
-			</div>
-		</template>
+		<SectionPropertiesPanel v-else-if="selected_section" />
 	</div>
 </template>
 
@@ -786,8 +605,8 @@ import TemplateInput from "./TemplateInput.vue";
 import LabelField from "./LabelField.vue";
 import SegmentedRow from "./SegmentedRow.vue";
 import StepperRow from "./StepperRow.vue";
-import PaddingGrid from "./PaddingGrid.vue";
 import Stepper from "./Stepper.vue";
+import SectionPropertiesPanel from "./SectionPropertiesPanel.vue";
 import { mountColorControl } from "./useColorControl";
 import { align_opts } from "./align_opts";
 
@@ -804,12 +623,6 @@ const open = ref({
 	f_field: true,
 	f_style: false,
 	f_visibility: false,
-	s_section: true,
-	s_bg: true,
-	s_padding: true,
-	s_layout: false,
-	s_style: false,
-	s_visibility: false,
 	t_table: true,
 	t_columns: true,
 	t_style: false,
@@ -1169,26 +982,6 @@ function set_image_size(col, value) {
 	col.image_size = isNaN(v) ? 40 : Math.max(16, Math.min(200, v));
 }
 
-// ── Section helpers ────────────────────────────────────────
-let section_orientation = computed(() => selected_section.value?.field_orientation ?? "");
-let section_gap = computed(() => selected_section.value?.gap ?? 20);
-const bg_color_host = ref(null);
-
-function mount_bg_color_control() {
-	mountColorControl(bg_color_host.value, {
-		value: selected_section.value?.background || "",
-		placeholder: __("Transparent"),
-		fieldname: "section_background",
-		onChange(value) {
-			if ((selected_section.value?.background ?? "") !== value) {
-				selected_section.value.background = value;
-			}
-		},
-	});
-}
-
-watch(selected_section, () => nextTick(mount_bg_color_control), { immediate: true });
-
 const rep_color_hosts = ref({});
 
 function mount_repeater_color_controls() {
@@ -1211,27 +1004,6 @@ watch(
 	() => nextTick(mount_repeater_color_controls),
 	{ immediate: true }
 );
-
-function set_columns(n) {
-	if (!selected_section.value) return;
-	const current = selected_section.value.columns.length;
-	if (n === current) return;
-	const all_fields = selected_section.value.columns.flatMap((col) => col.fields);
-	const new_columns = Array.from({ length: n }, () => ({ label: "", fields: [] }));
-	all_fields.forEach((field, i) => new_columns[i % n].fields.push(field));
-	selected_section.value.columns = new_columns;
-}
-
-let section_field_borders = computed(() => !!selected_section.value?.field_borders);
-let section_cell_padding = computed(() => selected_section.value?.cell_padding ?? 8);
-
-function toggle_field_borders(on) {
-	if (on) {
-		selected_section.value.field_borders = true;
-	} else {
-		delete selected_section.value.field_borders;
-	}
-}
 </script>
 
 <style scoped>
@@ -1339,61 +1111,6 @@ function toggle_field_borders(on) {
 	font-size: var(--text-sm);
 }
 
-/* ── Source display ──────────────────────────────────────── */
-.pfb-source-display {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding: 5px 8px;
-	border: 1px solid var(--border-color);
-	border-radius: var(--radius);
-	background: var(--control-bg);
-	gap: 6px;
-	min-width: 0;
-	overflow: hidden;
-}
-
-.pfb-source-name {
-	font-size: var(--text-sm);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	min-width: 0;
-}
-
-.pfb-type-badge {
-	font-size: var(--text-tiny);
-	color: var(--text-muted);
-	background: var(--gray-100);
-	border: 1px solid var(--gray-300);
-	border-radius: var(--radius);
-	padding: 1px 5px;
-	white-space: nowrap;
-	flex-shrink: 0;
-}
-
-/* ── Hint ────────────────────────────────────────────────── */
-
-/* ── Actions ─────────────────────────────────────────────── */
-
-.btn-danger-subtle {
-	display: inline-flex;
-	align-items: center;
-	gap: 4px;
-	color: var(--red-500);
-	background: transparent;
-	border: 1px solid var(--red-200);
-	border-radius: var(--radius);
-	padding: 5px 10px;
-	font-size: var(--text-sm);
-	cursor: pointer;
-}
-
-.btn-danger-subtle:hover {
-	background: var(--red-50);
-	border-color: var(--red-300);
-}
-
 /* ── Table column list ───────────────────────────────────── */
 .pfb-col-list {
 	padding: 4px 0;
@@ -1446,34 +1163,6 @@ function toggle_field_borders(on) {
 .pfb-col-label-input:focus {
 	border-color: var(--gray-500);
 	background: var(--fg-color);
-}
-
-.pfb-col-width-input {
-	width: 40px;
-	padding: 2px 4px;
-	font-size: var(--text-tiny);
-	text-align: right;
-	border: 1px solid var(--border-color);
-	border-radius: var(--radius);
-	background: var(--fg-color);
-	flex-shrink: 0;
-}
-
-.pfb-col-width-input:focus {
-	outline: none;
-	border-color: var(--gray-500);
-}
-
-/* hide number spin arrows */
-.pfb-col-width-input::-webkit-inner-spin-button,
-.pfb-col-width-input::-webkit-outer-spin-button {
-	-webkit-appearance: none;
-}
-
-.pfb-col-width-unit {
-	font-size: var(--text-tiny);
-	color: var(--text-muted);
-	flex-shrink: 0;
 }
 
 .pfb-col-remove {
