+++ frappe/public/js/print_format_builder/inspector.css
@@ -119,3 +119,71 @@
 	background: var(--subtle-accent);
 	color: var(--text-color);
 }
+.btn-danger-subtle {
+	display: inline-flex;
+	align-items: center;
+	gap: 4px;
+	color: var(--red-500);
+	background: transparent;
+	border: 1px solid var(--red-200);
+	border-radius: var(--radius);
+	padding: 5px 10px;
+	font-size: var(--text-sm);
+	cursor: pointer;
+}
+.btn-danger-subtle:hover {
+	background: var(--red-50);
+	border-color: var(--red-300);
+}
+.pfb-source-display {
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	padding: 5px 8px;
+	border: 1px solid var(--border-color);
+	border-radius: var(--radius);
+	background: var(--control-bg);
+	gap: 6px;
+	min-width: 0;
+	overflow: hidden;
+}
+.pfb-source-name {
+	font-size: var(--text-sm);
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+	min-width: 0;
+}
+.pfb-type-badge {
+	font-size: var(--text-tiny);
+	color: var(--text-muted);
+	background: var(--gray-100);
+	border: 1px solid var(--gray-300);
+	border-radius: var(--radius);
+	padding: 1px 5px;
+	white-space: nowrap;
+	flex-shrink: 0;
+}
+.pfb-col-width-input {
+	width: 40px;
+	padding: 2px 4px;
+	font-size: var(--text-tiny);
+	text-align: right;
+	border: 1px solid var(--border-color);
+	border-radius: var(--radius);
+	background: var(--fg-color);
+	flex-shrink: 0;
+}
+.pfb-col-width-input:focus {
+	outline: none;
+	border-color: var(--gray-500);
+}
+.pfb-col-width-input::-webkit-inner-spin-button,
+.pfb-col-width-input::-webkit-outer-spin-button {
+	-webkit-appearance: none;
+}
+.pfb-col-width-unit {
+	font-size: var(--text-tiny);
+	color: var(--text-muted);
+	flex-shrink: 0;
+}
